+++ provisioning.h
@@ -10,8 +10,8 @@ struct provisioning_sock {
 	void *udata;
 };
 
-int ksignal_defer_get_new_uuid(const char *base_url,
-                               struct provisioning_sock ps);
+intptr_t ksignal_defer_get_new_uuid(const char *base_url,
+                                    struct provisioning_sock ps);
 #define ksignal_defer_get_new_uuid(base_url, ...) \
 	ksignal_defer_get_new_uuid((base_url), \
 	                           (struct provisioning_sock){ __VA_ARGS__ })
